+++ server/component/BaseModel.php
@@ -232,6 +232,22 @@ abstract class BaseModel
         if($this->nav != null)
             return $this->nav->get_navigation_items();
         return array();
+    }    
+
+    /**
+     * Get a list of languages and prepares the list such that it can be passed to a
+     * list component.
+     *
+     * @retval array
+     *  An array of items where each item has the following keys:
+     *   'id':      The id of the language.
+     *   'locale':   
+     *   'language':   
+     *   'csv_separator':
+     */
+    public function get_languages()
+    {
+        return $this->db->get_languages();
     }
 }
 ?>