+++ CHANGELOG.md
@@ -14,6 +14,7 @@
     - `button`
     - `conditionalContainer`
  - load all custom css files in the mobile app
+ - set locale from mobile calls
 
 ### Changes
  - `formUserInput` can be added to a page with path containing `record_id`. Example: `/courseViewAdvanced/[i:record_id]?`. When the id is loaded the entry will be in edit mode. Also an option for delete will apear.